requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.12
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
typing-extensions==4.0.1
